Split UI library
