+++ LogcatCoreLib/src/main/AndroidManifest.xml
@@ -2,28 +2,7 @@
     package="info.hannes.logcat">
 
     <application>
-        <activity
-            android:name=".LogcatActivity"
-            android:label="Logcat"
-            android:theme="@style/DetailTheme" />
-        <activity
-            android:name=".LogfileActivity"
-            android:label="Timber"
-            android:theme="@style/DetailTheme" />
-        <activity
-            android:name=".BothLogActivity"
-            android:label="All log"
-            android:theme="@style/DetailTheme" />
 
-        <provider
-            android:name="androidx.core.content.FileProvider"
-            android:authorities="${applicationId}.provider"
-            android:exported="false"
-            android:grantUriPermissions="true">
-            <meta-data
-                android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/provider_paths"/>
-        </provider>
     </application>
 
 </manifest>
