+++ LogcatCoreUI/build.gradle
@@ -19,11 +19,13 @@ android {
 }
 
 dependencies {
+    api project(':LogcatCoreLib')
+    implementation "androidx.core:core-ktx:1.7.0"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1"
-    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.4.1"
-    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.4.1'
-    api 'com.jakewharton.timber:timber:5.0.1'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.5.2"
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.4.0"
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    api 'com.google.android.material:material:1.4.0'
 }
 
 project.afterEvaluate {
