+++ LogcatCoreUI/src/main/java/info/hannes/logcat/base/LogBaseFragment.kt
@@ -18,7 +18,7 @@ import androidx.lifecycle.coroutineScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.snackbar.Snackbar
-import info.hannes.logcat.R
+import info.hannes.logcat.ui.R
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
