+++ LogcatCoreUI/src/main/java/info/hannes/logcat/base/LogListAdapter.kt
@@ -9,7 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import info.hannes.logcat.R
+import info.hannes.logcat.ui.R
 import java.util.*
 
 class LogListAdapter(private var completeLogs: MutableList<String>, filter: String) : RecyclerView.Adapter<LogListAdapter.LogViewHolder>() {
