+++ LogcatCoreUI/src/main/java/info/hannes/logcat/ui/BothLogActivity.kt
@@ -1,11 +1,11 @@
-package info.hannes.logcat
+package info.hannes.logcat.ui
 
 import android.os.Bundle
 import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 
-class LogcatActivity : AppCompatActivity() {
+class BothLogActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -14,7 +14,6 @@ class LogcatActivity : AppCompatActivity() {
 
         // Check that the activity is using the layout version with the fragment_container FrameLayout
         if (findViewById<View>(R.id.fragment_container) != null) {
-
             // However, if we're being restored from a previous state,
             // then we don't need to do anything and should return or else
             // we could end up with overlapping fragments.
@@ -23,14 +22,19 @@ class LogcatActivity : AppCompatActivity() {
             }
 
             // Create a new Fragment to be placed in the activity layout
-            val firstFragment = LogcatFragment.newInstance("logcat.log", "search logcat")
+            val bothLogFragment = BothLogsFragment.newInstance(
+                "logfile.log",
+                "search logfile",
+                "search logcat",
+                "your@mail.com"
+            )
 
             // In case this activity was started with special instructions from an
             // Intent, pass the Intent's extras to the fragment as arguments
             // firstFragment.arguments = intent.extras
 
             // Add the fragment to the 'fragment_container' FrameLayout
-            supportFragmentManager.beginTransaction().add(R.id.fragment_container, firstFragment).commit()
+            supportFragmentManager.beginTransaction().add(R.id.fragment_container, bothLogFragment).commit()
         }
     }
