+++ LogcatCoreUI/src/main/java/info/hannes/logcat/ui/LogcatFragment.kt
@@ -1,4 +1,4 @@
-package info.hannes.logcat
+package info.hannes.logcat.ui
 
 import android.annotation.SuppressLint
 import android.os.Bundle
