+++ LogcatCoreUI/src/main/java/info/hannes/logcat/ui/LogfileActivity.kt
@@ -1,11 +1,11 @@
-package info.hannes.logcat
+package info.hannes.logcat.ui
 
 import android.os.Bundle
 import android.view.MenuItem
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 
-class BothLogActivity : AppCompatActivity() {
+class LogfileActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -14,6 +14,7 @@ class BothLogActivity : AppCompatActivity() {
 
         // Check that the activity is using the layout version with the fragment_container FrameLayout
         if (findViewById<View>(R.id.fragment_container) != null) {
+
             // However, if we're being restored from a previous state,
             // then we don't need to do anything and should return or else
             // we could end up with overlapping fragments.
@@ -22,11 +23,9 @@ class BothLogActivity : AppCompatActivity() {
             }
 
             // Create a new Fragment to be placed in the activity layout
-            val bothLogFragment = BothLogsFragment.newInstance(
-                    "logfile.log",
-                    "search logfile",
-                    "search logcat",
-                    "your@mail.com"
+            val firstFragment = LogfileFragment.newInstance(
+                "logfile.log",
+                "search logfile"
             )
 
             // In case this activity was started with special instructions from an
@@ -34,8 +33,9 @@ class BothLogActivity : AppCompatActivity() {
             // firstFragment.arguments = intent.extras
 
             // Add the fragment to the 'fragment_container' FrameLayout
-            supportFragmentManager.beginTransaction().add(R.id.fragment_container, bothLogFragment).commit()
+            supportFragmentManager.beginTransaction().add(R.id.fragment_container, firstFragment).commit()
         }
+
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
