+++ LogcatCoreUI/src/main/res/drawable/ic_menu_delete_24.xml
