+++ LogcatCoreUI/src/main/res/drawable/ic_menu_search_24.xml
