+++ LogcatCoreUI/src/main/res/drawable/ic_menu_share_24.xml
