+++ LogcatCoreUI/src/main/res/layout/content_log.xml
