+++ LogcatCoreUI/src/main/res/layout/item_log.xml
