+++ LogcatCoreUI/src/main/res/layout/loading_dialog.xml
