+++ LogcatCoreUI/src/main/res/values/attrs.xml
