+++ LogcatCoreUI/src/main/res/values/colors.xml
