+++ LogcatCoreUI/src/main/res/values/styles.xml
