+++ settings.gradle
@@ -1,4 +1,5 @@
 include ':LogcatCoreLib'
+include ':LogcatCoreUI'
 include ':LogcatCrashlyticLib'
 include ':LogcatCountlyLib'
 include ':sample'
